Rename interop client biary
