+++ interop/src/lib.rs
@@ -1,7 +1,7 @@
 #![recursion_limit = "256"]
 
 pub mod client;
-pub mod client_grpc;
+pub mod client_new_codegen;
 pub mod server;
 
 pub mod pb {
